Add 'always visible' channels to filtering mechanism
Previously it was only possible to mark channels that are
supposed to get filtered when channel filtering is active.
This patch adds the capability to mark channels that ought
to never be filtered also. This is important if the "Hide
empty channels when filtering" feature is used.

To enable this we add an additional column to the channels_filtered
table which tracks our new possible channel filter configuration
states:
* Always visible when filtering:
  The channels is always visible even if it is empty
  and filtering empty channels is enabled.
* Never visible when filtering:
  The channel is never visible. It is filtered even if it
  contains users. The only exception is if the user itself
  is in a channel in which case it is not filtered.
* Normal/Default:
  The channel is hidden when filtering if it is empty and
  hiding empty channels when filtering is enabled.

The current version of this patch uses the filter_on.svg
and filter_off.svg icons as placeholders. Instead two new
icons ought to be chosen for this purpose.

Issue: #1830